Transfer encryption methods to encrypt class.
