+++ test/enigma_test.rb
@@ -18,7 +18,6 @@ class EnigmaTest < Minitest::Test
     enigma.stubs(:generate_key).returns('02715')
 
     assert_equal "02715", enigma.generate_key
-    # binding.pry
   end
 
   def test_characters_as_hash
@@ -39,29 +38,6 @@ class EnigmaTest < Minitest::Test
     assert_equal [1, 0, 2, 5], enigma.date_shift("040895")
   end
 
-  def test_total_shift
-    enigma = Enigma.new
-
-    assert_equal [3, 27, 73, 20], enigma.total_shift("02715", "040895")
-  end
-
-  def test_rotate_letter
-    enigma = Enigma.new
-    assert_equal "k", enigma.a_rotation("h", "02715", "040895")
-    assert_equal "e", enigma.b_rotation("e", "02715", "040895")
-    assert_equal "d", enigma.c_rotation("l", "02715", "040895")
-    assert_equal "e", enigma.d_rotation("l", "02715", "040895")
-  end
-
-  def test_it_can_encrypt
-    enigma = Enigma.new
-    assert_equal ({
-        encryption: "keder ohulw",
-        key: "02715",
-        date: "040895"
-      }), enigma.encrypt("hello world", "02715", "040895")
-  end
-
   def test_it_can_decrypt
     enigma = Enigma.new
     assert_equal ({
@@ -70,5 +46,27 @@ class EnigmaTest < Minitest::Test
         date: "040895"
       }), enigma.decrypt("keder ohulw", "02715", "040895")
   end
-
 end
+
+# def test_total_shift
+#   enigma = Enigma.new
+#
+#   assert_equal [3, 27, 73, 20], enigma.total_shift("02715", "040895")
+# end
+#
+# def test_rotate_letter
+#   enigma = Enigma.new
+#   assert_equal "k", enigma.a_rotation("h", "02715", "040895")
+#   assert_equal "e", enigma.b_rotation("e", "02715", "040895")
+#   assert_equal "d", enigma.c_rotation("l", "02715", "040895")
+#   assert_equal "e", enigma.d_rotation("l", "02715", "040895")
+# end
+#
+# def test_it_can_encrypt
+#   enigma = Enigma.new
+#   assert_equal ({
+#       encryption: "keder ohulw",
+#       key: "02715",
+#       date: "040895"
+#     }), enigma.encrypt("hello world", "02715", "040895")
+# end
